Change about to TS
Also change from navlink to link and add some navigation types
