+++ src/App.jsx
@@ -3,8 +3,8 @@ import "./App.css";
 import Media from "./components/media/Media";
 import Navigation from "./components/Navigation";
 import Home from "./components/Home";
-import AboutALS from "./components/about/AboutALS";
-import AboutHT from "./components/about/AboutHT";
+import { AboutALS } from "./components/about/AboutALS";
+import { AboutHT } from "./components/about/AboutHT";
 import { NewsDetail} from "./components/news/NewsDetail";
 import Videos from "./components/media/Videos";
 import Photos from "./components/media/Photos";
