+++ src/components/about/Rules.tsx
@@ -1,0 +1,45 @@
+import React from "react";
+
+export const Rules: React.FC = () => {
+  return (
+    <div>
+      <h4>Simning</h4>
+      <p>
+        Det är frivilligt att använda våtdräkt oberoende av vattnets
+        temperatur. Alla måste använda den simmössa som delats ut vid
+        registreringen. Direkt efter simningen samlas simmössan in. Det är
+        inte tillåtet att använda fotklädnad i någon form eller handskar under
+        simningen. Följebåtar, surfingbrädor samt dykare finns runt hela
+        simsträckan. Får du problem kan du tillkalla hjälp från dom.
+      </p>
+      <h4>Cykling</h4>
+      <p>
+        Knäppt cykelhjälm ska bäras under hela cyklingen. Cykel måste ha
+        godkända bromsar. Inga hörlurar med musik är tillåtna. Elcykel är inte
+        tillåten. Cykling sker runt stora Hensjön, på både grus och asfalt.
+        Funktionärer finns vid varje vägkorsning. De vanliga trafikreglerna
+        gäller! På grussträckorna finns ”reservcyklar” utplacerade som går att
+        använda vid ev. haveri.
+      </p>
+      <h4>Löpning</h4>
+      <p>
+        Löpning sker på grus, gräs, skogsstig och en kort sträcka på asfalt.
+        Vägvakter finns vid varje vägövergång men det är alltid bilar som har
+        företräde. Det är alltså <strong>mycket viktigt</strong> att vara observant vid och över vägen. Inga hörlurar med musik
+        är tillåtna.
+      </p>
+
+      <h4>Generellt</h4>
+      <p>
+        <b>Du deltar på eget ansvar.</b>
+        <br></br>
+        Du har eget ansvar för att avgöra om du bör delta eller ej. Du har
+        eget ansvar för att din utrustning fungerar och uppfyller kraven. Du
+        har eget ansvar för din försäkring. Du har eget ansvar att läsa igenom
+        reglerna, och att agera efter dem och efter vanliga trafikregler.
+      </p>
+
+      Åldersgränsen är 16 år, men dispans kan ges ut vid förälders tillåtelse.
+    </div>
+  );
+}
